Cleaning and Exploring Data with Pandas
In this homework, we will investigate restaurant food safety scores for restaurants in San Francisco. The scores and violation information have been made available by the San Francisco Department of Public Health. The main goal for this assignment is to walk through the process of Data Cleaning and familiarise yourself with some of the pandas functions discussed in Pandas I and II.

After this homework, you should be comfortable with:

Reading CSV files,
Reading pandas documentation and using pandas,
Working with data at different levels of granularity, and
Identifying the type of data collected, missing values, anomalies, etc., and doing some basic analysis.