Food Safety II
In this homework, we will continue our exploration of restaurant food safety scores for restaurants in San Francisco. The main goal for this assignment is to focus more on the analysis of the dataset, building on the data cleaning we have done earlier in HW 2A.

After this homework, you should be comfortable with:

Reading pandas documentation and using pandas methods,
Working with data at different levels of granularity,
Using groupby with different aggregation functions, and
Chaining different pandas functions and methods to find answers to exploratory questions.